+++ src/threepp/math/vector2.cpp
@@ -6,9 +6,9 @@
 
 using namespace threepp;
 
-vector2::vector2(double x, double y) : x(x), y(y) {}
+vector2::vector2(float x, float y) : x(x), y(y) {}
 
-vector2 &vector2::set(double x, double y) {
+vector2 &vector2::set(float x, float y) {
 
     this->x = x;
     this->y = y;
@@ -16,7 +16,7 @@ vector2 &vector2::set(double x, double y) {
     return *this;
 }
 
-vector2 &vector2::setScalar(double value) {
+vector2 &vector2::setScalar(float value) {
 
     this->x = value;
     this->y = value;
@@ -24,21 +24,21 @@ vector2 &vector2::setScalar(double value) {
     return *this;
 }
 
-vector2 &vector2::setX(double value) {
+vector2 &vector2::setX(float value) {
 
     this->x = value;
 
     return *this;
 }
 
-vector2 &vector2::setY(double value) {
+vector2 &vector2::setY(float value) {
 
     y = value;
 
     return *this;
 }
 
-double &vector2::operator[](unsigned int index) {
+float &vector2::operator[](unsigned int index) {
     if (index >= 2) throw std::runtime_error("index out of bounds: " + std::to_string(index));
     switch (index) {
         case 0: return x;
@@ -55,7 +55,7 @@ vector2 &vector2::add(const vector2 &v) {
     return *this;
 }
 
-vector2 &vector2::add(double s) {
+vector2 &vector2::add(float s) {
 
     this->x += s;
     this->y += s;
@@ -71,7 +71,7 @@ vector2 &vector2::addVectors(const vector2 &a, const vector2 &b) {
     return *this;
 }
 
-vector2 &vector2::addScaledVector(const vector2 &v, double s) {
+vector2 &vector2::addScaledVector(const vector2 &v, float s) {
 
     this->x += v.x * s;
     this->y += v.y * s;
